refactor: add concept-form and collect all sub forms into this form, delete old forms
