+++ src/pages/concept-registration-page/form-allowed-and-discouraged-term/form-allowed-and-discouraged-term-pure.component.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Form, Field } from 'formik';
+import { Field } from 'formik';
 import { InputTagsField } from '../../../components/field-input-tags/field-input-tags.component';
 import { HelpText } from '../../../components/help-text/help-text.component';
 import { localization } from '../../../lib/localization';
 
 export const FormAllowedAndDiscouragedPure = (): JSX.Element => (
-  <Form>
+  <div>
     <div className="form-group">
       <HelpText
         title={localization.tillattTermTitle}
@@ -24,5 +24,5 @@ export const FormAllowedAndDiscouragedPure = (): JSX.Element => (
       />
       <Field name="frarådetTerm" component={InputTagsField} />
     </div>
-  </Form>
+  </div>
 );
