--- src/pages/concept-registration-page/form-allowed-and-discouraged-term/form-allowed-and-discouraged-term.schema.ts
@@ -1,7 +1,0 @@
-import * as Yup from 'yup';
-import { localization } from '../../../lib/localization';
-
-export const schema = Yup.object().shape({
-  tillattTerm: Yup.string().min(2, localization.validationMin2),
-  frarådetTerm: Yup.string().min(2, localization.validationMin2)
-});
