+++ src/pages/concept-registration-page/form-use-of-concept/form-useOfConcept-pure.component.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Form, Field } from 'formik';
+import { Field } from 'formik';
 import { InputField } from '../../../components/field-input/field-input.component';
 import { InputTagsField } from '../../../components/field-input-tags/field-input-tags.component';
 import { TextAreaField } from '../../../components/field-textarea/field-textarea.component';
@@ -7,7 +7,7 @@ import { HelpText } from '../../../components/help-text/help-text.component';
 import { localization } from '../../../lib/localization';
 
 export const FormUseOfTermPure = (): JSX.Element => (
-  <Form>
+  <div>
     <div className="form-group">
       <HelpText
         title={localization.eksempelTitle}
@@ -51,5 +51,5 @@ export const FormUseOfTermPure = (): JSX.Element => (
         </div>
       </div>
     </div>
-  </Form>
+  </div>
 );
